refactoring on_attribute with third size parameter

+++ test/datatypes.hpp
@@ -47,7 +47,7 @@ public:
     serializer.on_attribute("val_unsigned_long", unsigned_long_);
     serializer.on_attribute("val_unsigned_long_long", unsigned_long64_);
     serializer.on_attribute("val_bool", bool_);
-    serializer.on_attribute("val_cstr", cstr_, (size_t)CSTR_LEN);
+    serializer.on_attribute("val_cstr", cstr_, (long)CSTR_LEN);
     serializer.on_attribute("val_string", string_);
     serializer.on_attribute("val_varchar", varchar_, 63);
     serializer.on_attribute("val_date", date_);
